add module es6 function and webpck configuration
