--- assets/index.js
@@ -1,1 +1,0 @@
-import { Book } from "./book" ''
