+++ package.json
@@ -1,10 +1,10 @@
 {
   "name": "awesomebookses6",
   "version": "1.0.0",
-  "description": "<a name=\"readme-top\"></a>",
+  "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "webpack"
   },
   "keywords": [],
   "author": "",
@@ -17,5 +17,8 @@
     "stylelint-scss": "^3.21.0",
     "webpack": "^5.77.0",
     "webpack-cli": "^5.0.1"
+  },
+  "dependencies": {
+    "luxon": "^3.3.0"
   }
 }
