+++ src/assets/js/elements.js
@@ -1,0 +1,4 @@
+export const bookform = document.getElementById('form-asm-bookid');
+export const booklist = document.getElementById('awesomebookslist');
+export const booknavs = document.querySelectorAll('.nav-link');
+export const booknow = document.getElementById('section-date-now');
